refactor: move extractOverridDepsRec
