+++ lib/modules/manager/npm/extract/common/node.ts
@@ -1,0 +1,11 @@
+import type { PackageDependency } from '../../../types';
+import type { NpmManagerData } from '../../types';
+
+export function setNodeCommitTopic(
+  dep: PackageDependency<NpmManagerData>
+): void {
+  // This is a special case for Node.js to group it together with other managers
+  if (dep.depName === 'node') {
+    dep.commitMessageTopic = 'Node.js';
+  }
+}
